requirements.txt
Flask==3.0.0
loguru==0.7.2
python-dotenv==1.0.0
gunicorn==21.2.0
pyTelegramBotAPI==4.14.0
requests==2.31.0

